added the usage of AuthContex to app.jsx
app.jsx now has AuthContext Provider that gives the values for the context. the login function is used to login the user with either new expiry or old one depending on if the user logs in by themself or just refreshes the page. the useEffect with login as dep is used to keep people logged in even if they refresh the page. the expiration of the session is handled in the other useEffect that logs the user out if session has expired